El agregar de citas y reminders está completo. Notificaciones múltiples en Android funcionan.

+++ app/src/main/java/com/medbrain/fuzzy/medbrain/Appointment.java
@@ -149,7 +149,7 @@ public class Appointment {
      * @param int timeInMills
      * @return void
      */
-    public void setInnerCalendar(int timeInMills){
+    public void setInnerCalendar(long timeInMills){
         date.setTimeInMillis(timeInMills);
     }
